@@ -4,37 +4,36 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "@react-aria/utils": "^3.15.0",
-    "@react-oauth/google": "^0.9.0",
-    "@types/crypto-js": "^4.1.1",
-    "@types/jest": "^29.2.2",
-    "@types/node": "^18.11.9",
-    "@types/react": "^18.0.25",
-    "@types/react-big-calendar": "^0.24.7",
-    "@types/react-csv": "^1.1.1",
-    "@types/react-datepicker": "^3.1.5",
-    "@types/react-dom": "^18.0.8",
+    "@react-aria/utils": "^3.25.2",
+    "@react-oauth/google": "^0.12.1",
+    "@types/crypto-js": "^4.2.2",
+    "@types/jest": "^29.5.13",
+    "@types/node": "^22.5.5",
+    "@types/react": "^18.3.8",
+    "@types/react-big-calendar": "^1.8.11",
+    "@types/react-csv": "^1.1.10",
+    "@types/react-datepicker": "^7.0.0",
+    "@types/react-dom": "^18.3.0",
     "@types/react-router-dom": "^5.3.3",
-    "@types/react-router-hash-link": "^1.2.1",
-    "addresser": "^1.1.19",
-    "axios": "^1.6.0",
-    "classnames": "^2.3.2",
+    "@types/react-router-hash-link": "^2.4.9",
+    "addresser": "^1.1.20",
+    "axios": "^1.7.7",
+    "classnames": "^2.5.1",
     "crypto-js": "^4.2.0",
     "env-cmd": "^10.1.0",
-    "focus-trap-react": "^8.5.0",
-    "jwt-decode": "^3.1.2",
-    "moment": "^2.29.4",
-    "react": "^17.0.1",
-    "react-big-calendar": "^0.33.5",
-    "react-csv": "^2.0.3",
-    "react-datepicker": "^3.8.0",
-    "react-dom": "^17.0.2",
-    "react-hook-form": "^6.9.5",
-    "react-router-dom": "^5.3.4",
-    "react-router-hash-link": "^2.4.0",
+    "focus-trap-react": "^10.2.3",
+    "jwt-decode": "^4.0.0",
+    "moment": "^2.30.1",
+    "react": "^18.3.1",
+    "react-big-calendar": "^1.14.1",
+    "react-csv": "^2.2.2",
+    "react-datepicker": "^7.3.0",
+    "react-dom": "^18.3.1",
+    "react-hook-form": "^7.53.0",
+    "react-router-dom": "^6.26.2",
+    "react-router-hash-link": "^2.4.3",
     "react-scripts": "^5.0.1",
-    "reactjs-popup": "^2.0.4",
-    "typescript": "^4.8.4"
+    "reactjs-popup": "^2.0.6"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -66,6 +65,7 @@
     ]
   },
   "devDependencies": {
-    "eslint-config-prettier": "^8.3.0"
+    "eslint-config-prettier": "^9.1.0",
+    "typescript": "^4.9.5"
   }
 }
