@@ -4,20 +4,21 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "@typescript-eslint/parser": "^4.31.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "typescript": "^4.9.5"
+    "@typescript-eslint/parser": "^8.6.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-promise": "^7.1.0",
+    "react-scripts": "^5.0.1"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.8",
-    "@typescript-eslint/eslint-plugin": "^4.28.4",
+    "@babel/core": "^7.25.2",
+    "@typescript-eslint/eslint-plugin": "^8.6.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-import": "^2.23.4",
+    "eslint": "^8.57.1",
+    "eslint-plugin-import": "^2.30.0",
     "eslint-plugin-react": "^7.36.1",
-    "eslint-plugin-react-hooks": "^4.2.0",
-    "prettier": "2.8.8"
+    "eslint-plugin-react-hooks": "^4.6.2",
+    "prettier": "3.3.3",
+    "typescript": "^4.9.5"
   },
   "scripts": {
     "heroku-postbuild": "npm run build:frontend && npm run build:server",
