@@ -4,32 +4,32 @@
   "description": "Frontend code for Carriage",
   "main": "app.js",
   "dependencies": {
-    "@fast-csv/format": "^4.3.5",
-    "@types/cors": "^2.8.10",
-    "@types/express": "^4.17.8",
-    "@types/jsonwebtoken": "^8.5.0",
-    "@types/node-schedule": "^1.3.1",
-    "@types/uuid": "^8.3.0",
-    "@types/validator": "^13.1.3",
-    "@types/web-push": "^3.3.0",
-    "addresser": "^1.1.19",
-    "aws-sdk": "^2.1357.0",
-    "body-parser": "^1.20.1",
+    "@aws-sdk/client-dynamodb": "^3.656.0",
+    "@aws-sdk/client-s3": "^3.654.0",
+    "@aws-sdk/client-sns": "^3.654.0",
+    "@fast-csv/format": "^5.0.0",
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/jsonwebtoken": "^9.0.7",
+    "@types/node-schedule": "^2.1.7",
+    "@types/uuid": "^10.0.0",
+    "@types/validator": "^13.12.2",
+    "@types/web-push": "^3.6.3",
+    "addresser": "^1.1.20",
+    "body-parser": "^1.20.3",
     "cors": "^2.8.5",
-    "dotenv": "^8.2.0",
-    "dynamoose": "^2.7.3",
-    "express": "^4.18.2",
-    "google-auth-library": "^5.10.1",
-    "jsonwebtoken": "^9.0.0",
-    "moment": "^2.29.4",
-    "moment-timezone": "^0.5.35",
-    "node-schedule": "^2.0.0",
-    "nodemon": "^2.0.20",
-    "ts-node": "^8.10.2",
-    "typescript": "^4.0.3",
-    "uuid": "^8.3.1",
-    "validator": "^13.7.0",
-    "web-push": "^3.4.4"
+    "dotenv": "^16.4.5",
+    "dynamoose": "^4.0.1",
+    "express": "^4.21.0",
+    "google-auth-library": "^9.14.1",
+    "jsonwebtoken": "^9.0.2",
+    "moment": "^2.30.1",
+    "moment-timezone": "^0.5.45",
+    "node-schedule": "^2.1.1",
+    "nodemon": "^3.1.7",
+    "uuid": "^10.0.0",
+    "validator": "^13.12.0",
+    "web-push": "^3.6.7"
   },
   "scripts": {
     "build": "tsc --project tsconfig.build.json",
@@ -37,7 +37,7 @@
     "dev": "nodemon --exec \"ts-node\" src/app.ts",
     "lint": "eslint . --ext .js --ext .jsx --ext .ts --ext .tsx --fix",
     "type-check": "tsc --pretty --noEmit",
-    "test": "cross-env NODE_ENV=test mocha -r ts-node/register \"tests/**/*.ts\" --exit --timeout 10000",
+    "test": "cross-env NODE_ENV=test mocha --require ts-node/register --extensions ts,tsx --watch --watch-files src 'tests/**/*.{ts,tsx}' --exit --timeout 10000",
     "test:watch": "cross-env NODE_ENV=test nodemon --watch . --exec 'mocha -r ts-node/register \"tests/**/*.ts\" --timeout 10000' --ext ts"
   },
   "repository": {
@@ -51,12 +51,14 @@
   },
   "homepage": "https://github.com/cornell-dti/carriage-web#readme",
   "devDependencies": {
-    "@types/chai": "^4.3.4",
-    "@types/mocha": "^10.0.1",
-    "@types/supertest": "^2.0.12",
-    "chai": "^4.3.7",
+    "@types/chai": "^4.3.19",
+    "@types/mocha": "^10.0.8",
+    "@types/supertest": "^6.0.2",
+    "chai": "^4.5.0",
     "cross-env": "^7.0.3",
-    "mocha": "^10.2.0",
-    "supertest": "^6.3.3"
+    "mocha": "^10.7.3",
+    "supertest": "^7.0.0",
+    "ts-node": "^9.1.1",
+    "typescript": "^5.6.2"
   }
 }
