@@ -1,13 +1,13 @@
 import express from 'express';
 import { v4 as uuid } from 'uuid';
 import { Condition } from 'dynamoose';
-import { Document } from 'dynamoose/dist/Document';
 import moment from 'moment-timezone';
 import * as db from './common';
 import { Driver, DriverType, AvailabilityType } from '../models/driver';
 import { validateUser } from '../util';
 import { Ride, Status } from '../models/ride';
 import { UserType } from '../models/subscription';
+import { Item } from 'dynamoose/dist/Item';
 
 const router = express.Router();
 const tableName = 'Drivers';
@@ -108,7 +108,7 @@ router.get('/:id/stats', validateUser('Admin'), (req, res) => {
     }, 0);
 
   db.getById(res, Driver, id, tableName, (driver) => {
-    db.scan(res, Ride, condition, (data: Document[]) => {
+    db.scan(res, Ride, condition, (data: Item[]) => {
       res.send({
         data: {
           rides: data.length,
